refactor: share custom select ui (#392)
