+++ src/features/post-editor/ui/post-form.tsx
@@ -1,10 +1,6 @@
 "use client";
 
-import type {
-  FormEvent,
-  KeyboardEvent as ReactKeyboardEvent,
-  ReactNode,
-} from "react";
+import type { FormEvent, ReactNode } from "react";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { Icon } from "@iconify/react/offline";
 import archiveLinear from "@iconify-icons/solar/archive-linear";
@@ -60,7 +56,11 @@ import {
 import { normalizeAssetUrl, toCanonicalAssetUrl } from "@shared/lib/asset-url";
 import { getErrorMessage } from "@shared/lib/get-error-message";
 import { cn } from "@shared/lib/style-utils";
-import { Spinner } from "@shared/ui/libs";
+import {
+  CustomSelect,
+  Spinner,
+  type CustomSelectOption,
+} from "@shared/ui/libs";
 
 type EditorTab = "all" | "info" | "editor-split" | "editor-only";
 type SubmitIntent =
@@ -77,12 +77,6 @@ interface PostFormProps {
   cancelLabel?: string;
   onCancel?: () => void;
   onSuccess?: (post: PostDetail) => void;
-}
-
-interface InlineSelectOption<T extends string | number> {
-  label: string;
-  triggerLabel?: string;
-  value: T;
 }
 
 const DEFAULT_VALUES: PostFormValues = {
@@ -183,7 +177,7 @@ function sortCategories(categories: Category[]): Category[] {
 function flattenCategoryOptions(
   categories: Category[],
   depth = 0,
-): Array<{ label: string; triggerLabel?: string; value: number }> {
+): Array<CustomSelectOption<number>> {
   return categories.flatMap((category) => {
     const prefix = depth === 0 ? "" : `\u3000`.repeat(depth);
 
@@ -286,211 +280,6 @@ function SearchIndexableToggle({
         )}
       />
     </button>
-  );
-}
-
-function InlineCustomSelect<T extends string | number>({
-  value,
-  options,
-  onChange,
-  placeholder,
-  disabled,
-  className,
-}: {
-  value: T | null;
-  options: Array<InlineSelectOption<T>>;
-  onChange: (value: T) => void;
-  placeholder?: string;
-  disabled?: boolean;
-  className?: string;
-}) {
-  const [isOpen, setIsOpen] = useState(false);
-  const [activeIndex, setActiveIndex] = useState(-1);
-  const rootRef = useRef<HTMLDivElement | null>(null);
-  const listboxIdRef = useRef(
-    `inline-select-${Math.random().toString(36).slice(2)}`,
-  );
-  const optionRefs = useRef<Array<HTMLButtonElement | null>>([]);
-  const selected = options.find((option) => option.value === value);
-  const selectedIndex = options.findIndex((option) => option.value === value);
-
-  useEffect(() => {
-    optionRefs.current = [];
-  }, [options]);
-
-  useEffect(() => {
-    if (!isOpen) {
-      setActiveIndex(-1);
-
-      return;
-    }
-
-    setActiveIndex(selectedIndex >= 0 ? selectedIndex : 0);
-  }, [isOpen, selectedIndex]);
-
-  useEffect(() => {
-    if (!isOpen) {
-      return;
-    }
-
-    const handlePointerDown = (event: MouseEvent) => {
-      if (!rootRef.current?.contains(event.target as Node)) {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handlePointerDown);
-
-    return () => {
-      document.removeEventListener("mousedown", handlePointerDown);
-    };
-  }, [isOpen]);
-
-  useEffect(() => {
-    if (!isOpen || activeIndex < 0) {
-      return;
-    }
-
-    optionRefs.current[activeIndex]?.focus();
-  }, [activeIndex, isOpen]);
-
-  const commitSelection = (index: number) => {
-    const option = options[index];
-
-    if (!option) {
-      return;
-    }
-
-    onChange(option.value);
-    setIsOpen(false);
-  };
-
-  const handleTriggerKeyDown = (
-    event: ReactKeyboardEvent<HTMLButtonElement>,
-  ) => {
-    if (disabled) {
-      return;
-    }
-
-    if (event.key === "ArrowDown" || event.key === "ArrowUp") {
-      event.preventDefault();
-      setIsOpen(true);
-      setActiveIndex(selectedIndex >= 0 ? selectedIndex : 0);
-
-      return;
-    }
-
-    if (event.key === "Enter" || event.key === " ") {
-      event.preventDefault();
-      setIsOpen((current) => !current);
-    }
-  };
-
-  const handleOptionKeyDown = (
-    event: ReactKeyboardEvent<HTMLButtonElement>,
-    index: number,
-  ) => {
-    if (event.key === "ArrowDown") {
-      event.preventDefault();
-      setActiveIndex((index + 1) % options.length);
-
-      return;
-    }
-
-    if (event.key === "ArrowUp") {
-      event.preventDefault();
-      setActiveIndex((index - 1 + options.length) % options.length);
-
-      return;
-    }
-
-    if (event.key === "Home") {
-      event.preventDefault();
-      setActiveIndex(0);
-
-      return;
-    }
-
-    if (event.key === "End") {
-      event.preventDefault();
-      setActiveIndex(options.length - 1);
-
-      return;
-    }
-
-    if (event.key === "Escape") {
-      event.preventDefault();
-      setIsOpen(false);
-
-      return;
-    }
-
-    if (event.key === "Enter" || event.key === " ") {
-      event.preventDefault();
-      commitSelection(index);
-    }
-  };
-
-  return (
-    <div ref={rootRef} className={cn("relative w-full", className)}>
-      <button
-        type="button"
-        disabled={disabled}
-        onClick={() => setIsOpen((current) => !current)}
-        onKeyDown={handleTriggerKeyDown}
-        role="combobox"
-        aria-autocomplete="none"
-        aria-expanded={isOpen}
-        aria-controls={listboxIdRef.current}
-        aria-haspopup="listbox"
-        className="flex h-10 w-full items-center rounded-[0.5rem] border border-border-3 bg-background-1 px-3 pr-8 text-left text-[13px] text-text-2 outline-none transition-colors focus-visible:border-primary-1 disabled:cursor-not-allowed disabled:opacity-60"
-      >
-        <span className="truncate">
-          {selected?.triggerLabel ?? selected?.label ?? placeholder ?? ""}
-        </span>
-        <span className="pointer-events-none absolute right-3 top-1/2 -translate-y-1/2 text-[11px] text-text-4">
-          ▾
-        </span>
-      </button>
-
-      {isOpen ? (
-        <div
-          id={listboxIdRef.current}
-          role="listbox"
-          aria-activedescendant={
-            activeIndex >= 0
-              ? `${listboxIdRef.current}-option-${activeIndex}`
-              : undefined
-          }
-          className="absolute left-0 top-[calc(100%+0.25rem)] z-30 min-w-full overflow-hidden rounded-[0.5rem] border border-border-3 bg-background-1 shadow-[0px_8px_24px_rgba(15,23,42,0.08)]"
-        >
-          {options.map((option, index) => (
-            <button
-              key={String(option.value)}
-              id={`${listboxIdRef.current}-option-${index}`}
-              ref={(element) => {
-                optionRefs.current[index] = element;
-              }}
-              type="button"
-              role="option"
-              aria-selected={option.value === value}
-              tabIndex={activeIndex === index ? 0 : -1}
-              onKeyDown={(event) => handleOptionKeyDown(event, index)}
-              onFocus={() => setActiveIndex(index)}
-              onClick={() => commitSelection(index)}
-              className={cn(
-                "flex w-full items-center px-3 py-2 text-left text-[13px] transition-colors hover:bg-background-2",
-                option.value === value
-                  ? "font-medium text-primary-1"
-                  : "text-text-2",
-              )}
-            >
-              <span className="truncate">{option.label}</span>
-            </button>
-          ))}
-        </div>
-      ) : null}
-    </div>
   );
 }
 
@@ -1040,15 +829,19 @@ export function PostForm({
               <div className="flex flex-wrap items-center gap-4">
                 <CompactMetaLabel label="카테고리">
                   <div className="min-w-[10rem]">
-                    <InlineCustomSelect
+                    <CustomSelect
                       value={values.categoryId}
                       options={categoryOptions}
                       disabled={categoriesQuery.isPending}
+                      ariaLabel="카테고리"
                       placeholder={
                         categoriesQuery.isPending
                           ? "카테고리 불러오는 중..."
                           : "카테고리 선택"
                       }
+                      className="w-full"
+                      triggerClassName="rounded-[0.5rem] text-[13px] text-text-2 focus-visible:border-primary-1"
+                      optionClassName="text-[13px] text-text-2"
                       onChange={(value) =>
                         handleFieldChange("categoryId", value)
                       }
@@ -1161,9 +954,13 @@ export function PostForm({
 
                 <CompactMetaLabel label="댓글 상태">
                   <div className="min-w-[7rem]">
-                    <InlineCustomSelect
+                    <CustomSelect
                       value={values.commentStatus}
                       options={commentStatusOptions}
+                      ariaLabel="댓글 상태"
+                      className="w-full"
+                      triggerClassName="rounded-[0.5rem] text-[13px] text-text-2 focus-visible:border-primary-1"
+                      optionClassName="text-[13px] text-text-2"
                       onChange={(value) =>
                         handleFieldChange(
                           "commentStatus",
@@ -1182,15 +979,19 @@ export function PostForm({
               <div className="grid gap-6 lg:grid-cols-[minmax(0,1.15fr)_20rem]">
                 <div className="space-y-5">
                   <MetaFormRow label="카테고리">
-                    <InlineCustomSelect
+                    <CustomSelect
                       value={values.categoryId}
                       options={categoryOptions}
                       disabled={categoriesQuery.isPending}
+                      ariaLabel="카테고리"
                       placeholder={
                         categoriesQuery.isPending
                           ? "카테고리 불러오는 중..."
                           : "카테고리 선택"
                       }
+                      className="w-full"
+                      triggerClassName="rounded-[0.5rem] text-[13px] text-text-2 focus-visible:border-primary-1"
+                      optionClassName="text-[13px] text-text-2"
                       onChange={(value) =>
                         handleFieldChange("categoryId", value)
                       }
@@ -1255,9 +1056,13 @@ export function PostForm({
                     }: ${getCommentStatusDescription(values.commentStatus)}`}
                   >
                     <div className="max-w-[14rem]">
-                      <InlineCustomSelect
+                      <CustomSelect
                         value={values.commentStatus}
                         options={commentStatusOptions}
+                        ariaLabel="댓글 상태"
+                        className="w-full"
+                        triggerClassName="rounded-[0.5rem] text-[13px] text-text-2 focus-visible:border-primary-1"
+                        optionClassName="text-[13px] text-text-2"
                         onChange={(value) =>
                           handleFieldChange(
                             "commentStatus",
