+++ src/shared/ui/libs/index.tsx
@@ -1,6 +1,7 @@
 export { ArchiveHeader } from "./archive-header";
 export { ArchiveTagBadge } from "./archive-tag-badge";
 export { Button } from "./button";
+export { CustomSelect, type CustomSelectOption } from "./custom-select";
 export { EmptyState } from "./empty-state";
 export { ErrorContent } from "./error-content";
 export { ScrollToTop } from "./scroll-to-top";
