+++ src/widgets/admin-post-list/ui/post-filters.tsx
@@ -1,20 +1,13 @@
 "use client";
 
 import type { ReactNode } from "react";
-import {
-  type FormEvent,
-  useEffect,
-  useId,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { type FormEvent, useEffect, useMemo, useRef, useState } from "react";
 import { Icon } from "@iconify/react/offline";
-import altArrowDownLinear from "@iconify-icons/solar/alt-arrow-down-linear";
 import magniferLinear from "@iconify-icons/solar/magnifer-linear";
 import type { Category } from "@entities/category";
 import type { PostListItem } from "@entities/post";
 import { cn } from "@shared/lib/style-utils";
+import { CustomSelect } from "@shared/ui/libs";
 
 export type AdminPostTab = "active" | "trash";
 export type AdminPostStatusFilter = PostListItem["status"] | "all";
@@ -65,187 +58,6 @@ function flattenCategories(
     { category, depth },
     ...flattenCategories(categories, category.id, depth + 1),
   ]);
-}
-
-function FilterSelect({
-  value,
-  onChange,
-  className,
-  options,
-  ariaLabel,
-}: {
-  value: string;
-  onChange: (value: string) => void;
-  className?: string;
-  ariaLabel: string;
-  options: Array<{
-    label: string;
-    value: string;
-    depth?: number;
-  }>;
-}) {
-  const [open, setOpen] = useState(false);
-  const [activeIndex, setActiveIndex] = useState(0);
-  const rootRef = useRef<HTMLDivElement>(null);
-  const triggerRef = useRef<HTMLButtonElement>(null);
-  const optionRefs = useRef<Array<HTMLButtonElement | null>>([]);
-  const listboxId = useId();
-  const selectedOption =
-    options.find((option) => option.value === value) ?? options[0];
-  const selectedIndex = Math.max(
-    options.findIndex((option) => option.value === value),
-    0,
-  );
-
-  useEffect(() => {
-    function handlePointerDown(event: MouseEvent) {
-      if (rootRef.current && !rootRef.current.contains(event.target as Node)) {
-        setOpen(false);
-      }
-    }
-
-    document.addEventListener("mousedown", handlePointerDown);
-
-    return () => document.removeEventListener("mousedown", handlePointerDown);
-  }, []);
-
-  useEffect(() => {
-    if (!open) return;
-
-    const next = optionRefs.current[activeIndex];
-    if (!next) return;
-
-    const frame = requestAnimationFrame(() => next.focus());
-
-    return () => cancelAnimationFrame(frame);
-  }, [activeIndex, open]);
-
-  function openList(index = selectedIndex) {
-    setActiveIndex(index);
-    setOpen(true);
-  }
-
-  function closeList() {
-    setOpen(false);
-    triggerRef.current?.focus();
-  }
-
-  function selectValue(nextValue: string) {
-    onChange(nextValue);
-    closeList();
-  }
-
-  function moveActive(nextIndex: number) {
-    const maxIndex = options.length - 1;
-    setActiveIndex(Math.min(Math.max(nextIndex, 0), maxIndex));
-  }
-
-  return (
-    <div ref={rootRef} className={cn("relative inline-flex", className)}>
-      <button
-        ref={triggerRef}
-        type="button"
-        aria-label={ariaLabel}
-        aria-haspopup="listbox"
-        aria-expanded={open}
-        aria-controls={listboxId}
-        onClick={() => (open ? setOpen(false) : openList(selectedIndex))}
-        onKeyDown={(event) => {
-          if (event.key === "ArrowDown") {
-            event.preventDefault();
-            openList(selectedIndex);
-          } else if (event.key === "ArrowUp") {
-            event.preventDefault();
-            openList(selectedIndex);
-          } else if (event.key === "Enter" || event.key === " ") {
-            event.preventDefault();
-            openList(selectedIndex);
-          }
-        }}
-        className={cn(
-          "flex h-10 w-full items-center rounded-lg border border-border-3 bg-background-1 px-3 py-2 pr-8 text-left text-[14px] leading-5 text-text-1 outline-none transition-colors",
-          open && "border-primary-1 ring-3 ring-primary-1/10",
-        )}
-      >
-        <span className="truncate whitespace-nowrap">
-          {selectedOption?.label ?? ""}
-        </span>
-        <Icon
-          icon={altArrowDownLinear}
-          width="14"
-          aria-hidden="true"
-          className={cn(
-            "pointer-events-none absolute right-3 top-1/2 -translate-y-1/2 text-text-4 transition-transform",
-            open && "rotate-180",
-          )}
-        />
-      </button>
-
-      {open ? (
-        <div className="absolute left-0 top-[calc(100%+4px)] z-[200] min-w-full overflow-hidden rounded-lg border border-border-3 bg-background-1 shadow-[0_4px_16px_rgba(0,0,0,0.1)]">
-          <div
-            id={listboxId}
-            role="listbox"
-            aria-label={ariaLabel}
-            className="py-0.5"
-          >
-            {options.map((option, index) => {
-              const isSelected = option.value === value;
-
-              return (
-                <button
-                  key={option.value}
-                  ref={(node) => {
-                    optionRefs.current[index] = node;
-                  }}
-                  type="button"
-                  role="option"
-                  aria-selected={isSelected}
-                  tabIndex={index === activeIndex ? 0 : -1}
-                  onClick={() => selectValue(option.value)}
-                  onFocus={() => setActiveIndex(index)}
-                  onKeyDown={(event) => {
-                    if (event.key === "ArrowDown") {
-                      event.preventDefault();
-                      moveActive(index + 1);
-                    } else if (event.key === "ArrowUp") {
-                      event.preventDefault();
-                      moveActive(index - 1);
-                    } else if (event.key === "Home") {
-                      event.preventDefault();
-                      moveActive(0);
-                    } else if (event.key === "End") {
-                      event.preventDefault();
-                      moveActive(options.length - 1);
-                    } else if (event.key === "Escape") {
-                      event.preventDefault();
-                      closeList();
-                    } else if (event.key === "Tab") {
-                      setOpen(false);
-                    } else if (event.key === " " || event.key === "Enter") {
-                      event.preventDefault();
-                      selectValue(option.value);
-                    }
-                  }}
-                  className={cn(
-                    "flex w-full items-center whitespace-nowrap px-3 py-2 text-left text-[14px] leading-5 text-text-1 transition-colors hover:bg-background-2",
-                    isSelected && "font-medium text-primary-1",
-                  )}
-                  style={{
-                    paddingLeft: option.depth
-                      ? `${12 + option.depth * 16}px`
-                      : undefined,
-                  }}
-                >
-                  {option.label}
-                </button>
-              );
-            })}
-          </div>
-        </div>
-      ) : null}
-    </div>
-  );
 }
 
 export function PostFilters({
@@ -341,7 +153,7 @@ export function PostFilters({
           onSubmit={handleSearch}
           className="flex flex-wrap items-center gap-3"
         >
-          <FilterSelect
+          <CustomSelect
             value={status}
             onChange={(value) => onStatusChange(value as AdminPostStatusFilter)}
             className="min-w-[8.5rem]"
@@ -349,7 +161,7 @@ export function PostFilters({
             options={STATUS_OPTIONS}
           />
 
-          <FilterSelect
+          <CustomSelect
             value={visibility}
             onChange={(value) =>
               onVisibilityChange(value as AdminPostVisibilityFilter)
@@ -359,7 +171,7 @@ export function PostFilters({
             options={VISIBILITY_OPTIONS}
           />
 
-          <FilterSelect
+          <CustomSelect
             value={categoryId ? String(categoryId) : ""}
             onChange={(value) =>
               onCategoryChange(value ? Number(value) : undefined)
